% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ML_parameters_scale_free_confounders}
\alias{get_ML_parameters_scale_free_confounders}
\title{Derive structural parameters at ML solution given scale-free confounders.}
\usage{
get_ML_parameters_scale_free_confounders(Sigma_hat, sc24, sc34)
}
\arguments{
\item{Sigma_hat}{sample covariance matrix}

\item{sc24}{scaled confounding coefficient for 4 -> 2}

\item{sc34}{scaled confounding coefficient for 4 -> 3}
}
\value{
List containing matrix B (structural parameters in original scale)
and diagonal matrix V (intrinsic variances).
}
\description{
Derive structural parameters at ML solution given scale-free confounders.
}
\examples{
Sigma <- get_Sigma(1, 0.05, 1, 0, 0, 2, 3, 4)
get_ML_parameters_scale_free_confounders(Sigma, 0, 0)
}
