% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Sigma}
\alias{get_Sigma}
\title{Derive covariance matrix Sigma from model parameters}
\usage{
get_Sigma(b21, b31, b32, c24, c34, v1, v2, v3)
}
\arguments{
\item{b21}{structural coefficient for 1 -> 2}

\item{b31}{structural coefficient for 1 -> 3}

\item{b32}{structural coefficient for 2 -> 3}

\item{c24}{confounding coefficient for 4 -> 2}

\item{c34}{confounding coefficient for 4 -> 3}

\item{v1}{intrinsic variance of 1}

\item{v2}{intrinsic variance of 2}

\item{v3}{intrinsic variance of 3}
}
\value{
model covariance matrix given the path coefficients and variances
}
\description{
The function covariance matrix for the linear-Gaussian model
given all the parameter values.
}
\examples{
get_Sigma(1, 0.05, 1, 0, 0, 1, 1, 1)
}
