% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illustrative_example_MultiNest_samples}
\alias{illustrative_example_MultiNest_samples}
\title{A dataset containing simulated samples for reproducing Figure 3
Section 4 of the 2017 AISTATS paper.}
\format{
A list containing six coda::mcmc objects, which represent the
posterior samples for each scenario.
}
\usage{
illustrative_example_MultiNest_samples
}
\description{
A dataset containing simulated samples for reproducing Figure 3
Section 4 of the 2017 AISTATS paper.
}
\keyword{datasets}
