% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_functions.R
\name{log_hessian_term_ML_manifold}
\alias{log_hessian_term_ML_manifold}
\title{Function to derive the RoCELL log-determinant of the (negative) Hessian on
the ML manifold, where for each pair of confounding coefficients we can find
a unique ML solution.}
\usage{
log_hessian_term_ML_manifold(sc24, sc34, Sigma_hat, N)
}
\arguments{
\item{sc24}{confounding coefficient from 2 to 4 (scale-free)}

\item{sc34}{confounding coefficient from 3 to 4 (scale-free)}

\item{Sigma_hat}{sample covariance matrix}

\item{N}{number of samples}
}
\value{
the (negative) Hessian log-determinant
}
\description{
Function to derive the RoCELL log-determinant of the (negative) Hessian on
the ML manifold, where for each pair of confounding coefficients we can find
a unique ML solution.
}
\examples{
Sigma <- get_Sigma(1, 0.05, 1, 1, 1, 1, 1, 1)
log_hessian_term_ML_manifold(0, 0, Sigma, 1)
}
