% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_functions.R
\name{log_posterior_ML_manifold}
\alias{log_posterior_ML_manifold}
\title{Function to derive the RoCELL log-posterior on the ML manifold, where for
each pair of confounding coefficients we can find a unique ML solution.}
\usage{
log_posterior_ML_manifold(sc24, sc34, Sigma_hat, N, k_slab = 1, k_spike = 100)
}
\arguments{
\item{sc24}{confounding coefficient from 2 to 4 (scale-free)}

\item{sc34}{confounding coefficient from 3 to 4 (scale-free)}

\item{Sigma_hat}{sample covariance matrix}

\item{N}{number of samples}

\item{k_slab}{prior slab component precision}

\item{k_spike}{prior spike component precision}
}
\value{
log-posterior value
}
\description{
Function to derive the RoCELL log-posterior on the ML manifold, where for
each pair of confounding coefficients we can find a unique ML solution.
}
\examples{
Sigma <- get_Sigma(1, 0.05, 1, 1, 1, 1, 1, 1)
log_posterior_ML_manifold(0, 0, Sigma, 1)
}
