% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_selection_MultiNest_log_evidences}
\alias{model_selection_MultiNest_log_evidences}
\title{A dataset containing simulated samples for reproducing Figure 7
Section 4 of the 2017 AISTATS paper.}
\format{
A data frame containing the estimated log-evidences and standard errors.
\describe{
\item{Spike}{spike component precision}
\item{Correct}{Log-evidence in the correct direction (of the generating model)}
\item{Correct Error}{Standard error for estimated log-evidence}
\item{Reverse}{Log-evidence in the reverse direction (of the generating model)}
\item{Reverse Error}{Standard error for estimated log-evidence}
}
}
\usage{
model_selection_MultiNest_log_evidences
}
\description{
A dataset containing simulated samples for reproducing Figure 7
Section 4 of the 2017 AISTATS paper.
}
\keyword{datasets}
