% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_get_params_scale_free}
\alias{Rcpp_get_params_scale_free}
\title{Derive scale-free structural parameters at ML solution given scale-free confounding coefficients.}
\usage{
Rcpp_get_params_scale_free(Sigma_hat, sc24, sc34)
}
\arguments{
\item{Sigma_hat}{sample covariance matrix}

\item{sc24}{scaled confounding coefficient for 4 -> 2}

\item{sc34}{scaled confounding coefficient for 4 -> 3}
}
\value{
Vector containing scaled structural coefficients (sb21, sb31, sb32)
and intrinsic variances (v1, v2, v3), in this order.
}
\description{
Derive scale-free structural parameters at ML solution given scale-free confounding coefficients.
}
