% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_log_likelihood_hessian_scale_free}
\alias{Rcpp_log_likelihood_hessian_scale_free}
\title{Compute log-likelihood Hessian wrt scale-free parameters.}
\usage{
Rcpp_log_likelihood_hessian_scale_free(
  Sigma_hat,
  N,
  sb21,
  sb31,
  sb32,
  sc24,
  sc34,
  v1,
  v2,
  v3
)
}
\arguments{
\item{Sigma_hat}{sample covariance matrix}

\item{N}{number of samples}

\item{sb21}{scaled structural coefficient for 1 -> 2}

\item{sb31}{scaled structural coefficient for 1 -> 3}

\item{sb32}{scaled structural coefficient for 1 -> 3}

\item{sc24}{scaled confounding coefficient for 4 -> 2}

\item{sc34}{scaled confounding coefficient for 4 -> 3}

\item{v1}{intrinsic variance of 1}

\item{v2}{intrinsic variance of 2}

\item{v3}{intrinsic variance of 3}
}
\value{
log-likelihood Hessian matrix (8x8) with the parameters ordered as above
}
\description{
Compute log-likelihood Hessian wrt scale-free parameters.
}
