% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_log_posterior_RoCELL}
\alias{Rcpp_log_posterior_RoCELL}
\title{Compute RoCELL log-posterior given sample covariance matrix and parameters.}
\usage{
Rcpp_log_posterior_RoCELL(
  Sigma_hat,
  N,
  sb21,
  sb31,
  sb32,
  sc24,
  sc34,
  v1,
  v2,
  v3,
  slab_precision,
  spike_precision,
  w32 = 0.5
)
}
\arguments{
\item{Sigma_hat}{sample covariance matrix}

\item{N}{number of samples}

\item{sb21}{scaled structural coefficient for 1 -> 2}

\item{sb31}{scaled structural coefficient for 1 -> 3}

\item{sb32}{scaled structural coefficient for 1 -> 3}

\item{sc24}{scaled confounding coefficient for 4 -> 2}

\item{sc34}{scaled confounding coefficient for 4 -> 3}

\item{v1}{intrinsic variance of 1}

\item{v2}{intrinsic variance of 2}

\item{v3}{intrinsic variance of 3}

\item{slab_precision}{slab component precision}

\item{spike_precision}{spike component precision}

\item{w32}{Mixture weight for spike-and-slab prior applied to sb32.}
}
\value{
RoCELL log-posterior computed at the specified parameter values
}
\description{
Compute RoCELL log-posterior given sample covariance matrix and parameters.
}
