% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_log_spike_and_slab_scale_free}
\alias{Rcpp_log_spike_and_slab_scale_free}
\title{Compute log-value of spike-and-slab prior for parameter given mixture weight.}
\usage{
Rcpp_log_spike_and_slab_scale_free(x, w, k_slab = 1, k_spike = 100)
}
\arguments{
\item{x}{numeric value of random variable}

\item{w}{mixture weight between 0 and 1, with 1 indicating slab and 0 spike}

\item{k_slab}{slab component precision}

\item{k_spike}{spike component precision}
}
\value{
Log-value of spike-and-slab density computed at x.
}
\description{
Compute log-value of spike-and-slab prior for parameter given mixture weight.
}
