% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_MultiNest_log_evidence}
\alias{read_MultiNest_log_evidence}
\title{Read log evidence from MultiNest stats output file}
\usage{
read_MultiNest_log_evidence(filename = "output_RoCELL-stats.dat")
}
\arguments{
\item{filename}{character; Name of MultiNest stats.dat output file.}
}
\value{
Numeric vector containing the estimated log-evidence as the first
element and its estimated standard error as the second element.
}
\description{
Read log evidence from MultiNest stats output file
}
\examples{
\dontrun{
Sigma <- get_Sigma(1, 0.05, 1, 1, 1, 1, 1, 1)
run_system_command(sprintf(
  "./RoCELL -s11 \%g -s12 \%g -s13 \%g -s22 \%g -s23 \%g -s33 \%g",
  Sigma[1, 1], Sigma[1, 2], Sigma[1, 3], Sigma[2, 2], Sigma[2, 3], Sigma[3, 3]
))
read_MultiNest_log_evidence("output_RoCELL-stats.dat")
}
}
