% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_MultiNest_samples}
\alias{read_MultiNest_samples}
\title{Read RoCELL MultiNest output given sample covariance matrix used as input.}
\usage{
read_MultiNest_samples(filename, Sigma_hat)
}
\arguments{
\item{filename}{character; Name of file containing RoCELL MultiNest output (.txt).}

\item{Sigma_hat}{sample covariance matrix}
}
\value{
coda::mcmc structure containing the RoCELL posterior samples
}
\description{
Read RoCELL MultiNest output given sample covariance matrix used as input.
}
\examples{
\dontrun{
Sigma <- get_Sigma(1, 0.05, 1, 1, 1, 1, 1, 1)
run_system_command(sprintf(
  "./RoCELL -s11 \%g -s12 \%g -s13 \%g -s22 \%g -s23 \%g -s33 \%g",
  Sigma[1, 1], Sigma[1, 2], Sigma[1, 3], Sigma[2, 2], Sigma[2, 3], Sigma[3, 3]
))
read_MultiNest_samples("output_RoCELL-.txt", Sigma)
}
}
