% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_variance_robustness_MultiNest_samples}
\alias{spike_variance_robustness_MultiNest_samples}
\title{A dataset containing simulated samples for reproducing Figure 8 in
Section 4 of the 2017 AISTATS paper.}
\format{
A list containing ten coda::mcmc objects, which represent the
posterior samples for each spike width (variance).
}
\usage{
spike_variance_robustness_MultiNest_samples
}
\description{
A dataset containing simulated samples for reproducing Figure 8 in
Section 4 of the 2017 AISTATS paper.
}
\keyword{datasets}
